feat: Implement logic component based on pyrenees page
